refactor: reorganize components by domain
