+++ web/app/baskets/[id]/work/page.tsx
@@ -1,5 +1,5 @@
-import BasketWorkbenchLayout from "@/components/basket-work/BasketWorkbenchLayout";
-import ContextBlocksPanel from "@/components/basket-work/ContextBlocksPanel";
+import BasketWorkbenchLayout from "@/components/basket/BasketWorkbenchLayout";
+import ContextBlocksPanel from "@/components/basket/ContextBlocksPanel";
 import { createServerSupabaseClient } from "@/lib/supabaseServerClient";
 import { redirect } from "next/navigation";
