+++ web/app/blocks/layout.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { ReactNode } from "react";
-import Shell from "@/components/layouts/Shell";
+import Shell from "@/components/layout/Shell";
 
 export default function Layout({ children }: { children: ReactNode }) {
     return <Shell>{children}</Shell>;
