+++ web/app/integrations/layout.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { ReactNode } from "react";
-import Shell from "@/components/layouts/Shell";
+import Shell from "@/components/layout/Shell";
 
 export default function IntegrationsLayout({ children }: { children: ReactNode }) {
   return <Shell>{children}</Shell>;
