+++ web/components/basket/ContextBlocksPanel.tsx
