+++ web/components/briefs/index.ts
@@ -1,0 +1,1 @@
+export {}
