+++ web/components/common/RightPanel.tsx
