+++ web/components/common/UploadArea.tsx
